requirements.txt
alembic==1.14.0
aniso8601==9.0.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
PyMySQL==1.1.1
pytz==2024.2
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
